requirements.txt
absl-py==0.12.0
aiohttp==3.8.1
aioice==0.7.5
aiortc==1.2.1
aiosignal==1.2.0
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
av==8.1.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
cachetools==4.2.4
certifi==2018.1.18
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==6.7
colorama==0.3.7
cryptography==36.0.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.26
dataclasses==0.8
decorator==4.1.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dnspython==2.2.0rc1
feedparser==5.2.1
flatbuffers==1.12
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-crc32c==1.3.0
google-pasta==0.2.0
graphsurgeon==0.4.5
grpcio==1.34.1
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
idna-ssl==1.1.0
importlib-metadata==4.8.3
jetson-stats==3.1.1
Jetson.GPIO==2.0.17
joblib==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.6
MarkupSafe==1.0
matplotlib==2.1.1
mediapipe @ file:///home/dmp/1.Users/1.TinhLam/2.Build_mediapipe/mediapipe-bin/mediapipe-0.8.5_cuda102-cp36-none-linux_aarch64.whl
mock==3.0.5
multidict==5.2.0
netifaces==0.11.0
numpy==1.19.4
oauth==1.0.1
oauthlib==3.1.1
onboard==1.4.1
opencv-contrib-python==4.5.5.62
opt-einsum==3.3.0
PAM==0.4.2
pandas==0.22.0
pbr==5.8.0
pkgconfig==1.5.5
protobuf==3.19.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.0
pycairo==1.16.2
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pyee==8.2.2
PyGObject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pylibsrtp==0.7.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.8
scikit-learn==0.21.3
scipy==1.5.4
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
sklearn==0.0
slidingwindow==0.0.14
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0+nv21.8
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorrt==8.0.1.6
termcolor==1.1.0
testresources==2.0.1
threadpoolctl==3.0.0
tqdm==4.62.3
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
uff==0.6.9
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
Werkzeug==2.0.2
wrapt==1.12.1
xkit==0.0.0
yarl==1.7.2
youtube_dl==2018.3.14
zipp==3.6.0
zope.interface==4.3.2
